Refactor touchtechnology.common.forms
