+++ touchtechnology/common/db/models.py
@@ -1,28 +1,13 @@
 import uuid
 
 from django.db import models
-from django.db.models import Min
-from django.utils.encoding import smart_str
-from mptt.forms import TreeNodeChoiceField
-from touchtechnology.common.forms import (
+from touchtechnology.common.forms.fields import (
     BooleanChoiceField, EmailField as EmailFormField, GoogleMapsField,
-    HTMLField as HTMLFormField, ModelChoiceField, SelectDateField,
-    SelectDateTimeField, SelectTimeField, SitemapNodeModelChoiceField,
-    TemplatePathFormField, iCheckModelMultipleChoiceField,
+    HTMLField as HTMLFormField, MinTreeNodeChoiceField, ModelChoiceField,
+    SelectDateField, SelectDateTimeField, SelectTimeField,
+    SitemapNodeModelChoiceField, TemplatePathFormField,
+    iCheckModelMultipleChoiceField,
 )
-
-
-class MinTreeNodeChoiceField(TreeNodeChoiceField):
-    @property
-    def minimum_level(self):
-        if not hasattr(self, '_minimum_level'):
-            self._minimum_level = self.queryset.aggregate(
-                minimum=Min('level')).get('minimum')
-        return self._minimum_level
-
-    def label_from_instance(self, obj):
-        level = obj.level - self.minimum_level
-        return u'%s %s' % (self.level_indicator * level, smart_str(obj))
 
 
 class TreeField(models.ForeignKey):
