+++ touchtechnology/common/views.py
@@ -4,7 +4,7 @@ from django.shortcuts import redirect
 from django.utils.http import is_safe_url, urlencode
 from django.utils.six.moves.urllib.parse import parse_qsl, urlsplit, urlunsplit
 from django.views.decorators.http import require_POST
-from touchtechnology.common.forms import SelectTimezoneForm
+from touchtechnology.common.forms.tz import SelectTimezoneForm
 
 
 def login(request, to, *args, **kwargs):
