+++ touchtechnology/news/forms.py
@@ -5,8 +5,8 @@ from django.conf import settings
 from django.forms.models import inlineformset_factory
 from django.utils.translation import ugettext_lazy as _
 from modelforms.forms import ModelForm
-from touchtechnology.common.forms import SuperUserSlugMixin
-from touchtechnology.common.mixins import BootstrapFormControlMixin
+from touchtechnology.common.forms import BootstrapFormControlMixin
+from touchtechnology.common.forms.mixins import SuperUserSlugMixin
 from touchtechnology.content.forms import PlaceholderConfigurationBase
 from touchtechnology.news.app_settings import NEWS_CONTENT_BLOCKS
 from touchtechnology.news.models import Article, ArticleContent, Category
